requirements.txt
numpy==1.23.5
opencv_python==4.7.0.72
pandas==1.5.3
Pillow==9.5.0
streamlit==1.20.0
